state: moved stdlib lock to upper function to fix race condition
